you know x32 isn't a thing, it's more like x86. Also your thing is broken
long description : 
intel made the first microprocessor called the 4004 and was a 4 bit CPU. then they made the first 8 bit CPU called the 8008, which then became the 8080. In 1978, intel made the 8086, a 16-bit CPU which would then become the 8088 for the first IBM PC in 1981. That CPU and consequent were based on the 8086, and the architecture was called x86. Many computers apart from Apple Macs and some small computer brands used x86, and it's only now that Apple tried ARM(Apple Silicon) chips. Even then, most computers still use x86.