THIS IS SAFE AS I HAVE TESTED THIS ON MY COMPUTER.
I have tested and used it on my daily driver PC and none of my accounts have gotten hacked, nor has my data been compromised. Check in the top-right corner if the indicator is green. If it is not, that version is unsafe. We are not responsible to what happens to your computer. It can blow up and we would not be liable for all of the damages. 

GAMERSOFT DO NOT CHANGE THIS OR IF YOU CHANGE IT USE THE SAME COMMIT MESSAGES.